server/requirements.txt
aiogram
psycopg2-binary
python-dotenv
rq
Pillow
requests